Parse query vars containing equals sign (=)
When a query variable contains an equals sign, the QueryString
parsing will strip the equals sign. One example of would be with
base64 encoded strings containing an equals sign at the end for
padding. While this issue can be prevented by simply urlencoding
the URL, cURL, for example, handles this gracefully.